+++ realm/realm-library/src/main/java/io/realm/internal/Context.java
@@ -133,14 +133,6 @@ public class Context {
         }
     }
 
-    public void asyncDisposeGroup(long nativePointer) {
-        Group.nativeClose(nativePointer);
-    }
-
-    public void asyncDisposeSharedGroup(long nativePointer) {
-        SharedGroup.nativeClose(nativePointer);
-    }
-
     protected void finalize() throws Throwable {
         synchronized (this) {
             isFinalized = true;